test.requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
certifi==2025.1.31
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.3
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
